issuer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
jwcrypto==1.5.6
pydantic==2.9.0
